Add `server-framework-tonic` crate
