+++ Cargo.toml
@@ -55,4 +55,7 @@ serde_json = "1.0"
 tracing-subscriber = { version = "0.3", features = ["json", "env-filter"] }
 
 [workspace]
-members = ["examples/*"]
+members = [
+    "server-framework-tonic",
+    "examples/*"
+]
