+++ examples/example-tonic/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 [dependencies]
 prost = "0.9"
 server-framework = { path = "../../" }
+server-framework-tonic = { path = "../../server-framework-tonic" }
 tokio = { version = "1.14", features = ["full"] }
 tonic = "0.6"
 tracing = "0.1"
