+++ src/server.rs
@@ -188,69 +188,6 @@ impl<F, H> Server<F, H> {
         self
     }
 
-    /// Add a tonic service to the server.
-    ///
-    /// ```rust
-    /// use axum::async_trait;
-    /// use server_framework::Server;
-    /// #
-    /// # #[async_trait]
-    /// # trait Greeter {}
-    /// # #[derive(Clone)]
-    /// # struct GreeterServer<T>(T);
-    /// # impl<T> GreeterServer<T> {
-    /// #     fn new(t: T) -> Self { Self(t) }
-    /// # }
-    /// # impl<T> tonic::transport::NamedService for GreeterServer<T> {
-    /// #     const NAME: &'static str = "";
-    /// # }
-    /// # impl<T> tower::Service<http::Request<axum::body::BoxBody>> for GreeterServer<T> {
-    /// #     type Response = axum::response::Response;
-    /// #     type Error = tonic::codegen::Never;
-    /// #     type Future = std::future::Ready<Result<Self::Response, Self::Error>>;
-    /// #     fn poll_ready(&mut self, _: &mut std::task::Context<'_>) -> std::task::Poll<Result<(), Self::Error>> {
-    /// #         todo!()
-    /// #     }
-    /// #     fn call(&mut self, _: http::Request<axum::body::BoxBody>) -> Self::Future {
-    /// #         todo!()
-    /// #     }
-    /// # }
-    ///
-    /// #[derive(Clone)]
-    /// struct MyGreeter;
-    ///
-    /// // implement server trait generated by tonic-build
-    /// #[async_trait]
-    /// impl Greeter for MyGreeter {
-    ///     // ...
-    /// }
-    ///
-    /// let service = GreeterServer::new(MyGreeter);
-    ///
-    /// # async {
-    /// Server::default()
-    ///     .with_tonic(service)
-    ///     .always_live_and_ready()
-    ///     .serve()
-    ///     .await
-    ///     .unwrap();
-    /// # };
-    /// ```
-    #[cfg(feature = "tonic")]
-    pub fn with_tonic<S, B>(self, service: S) -> Self
-    where
-        S: Service<Request, Response = Response<B>, Error = tonic::codegen::Never>
-            + tonic::transport::NamedService
-            + Clone
-            + Send
-            + 'static,
-        S::Future: Send,
-        B: http_body::Body<Data = axum::body::Bytes> + Send + 'static,
-        B::Error: Into<axum::BoxError>,
-    {
-        self.with(router_from_tonic(service))
-    }
-
     /// Router all requests to the given service.
     ///
     /// Note that _all_ requests will be sent to the service and therefore the server cannot
@@ -660,26 +597,4 @@ async fn signal_listener() {
         .await
         .expect("Failed to listen for Ctrl-c signal");
     tracing::info!("Ctrl-c received, shutting down server");
-}
-
-/// Convert a [`tonic`] service into a [`Router`].
-///
-/// This can be useful for composing a number of services and adding middleware to them.
-#[cfg(feature = "tonic")]
-pub fn router_from_tonic<S, B>(service: S) -> Router<BoxBody>
-where
-    S: Service<Request, Response = Response<B>, Error = tonic::codegen::Never>
-        + tonic::transport::NamedService
-        + Clone
-        + Send
-        + 'static,
-    S::Future: Send,
-    B: http_body::Body<Data = axum::body::Bytes> + Send + 'static,
-    B::Error: Into<axum::BoxError>,
-{
-    let svc = ServiceBuilder::new()
-        .map_err(|err: tonic::codegen::Never| match err {})
-        .map_response_body(body::boxed)
-        .service(service);
-    Router::new().route(&format!("/{}/*rest", S::NAME), svc)
 }
